requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.5
python-dotenv>=1.0
aiohttp>=3.9,<4
msal>=1.28
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
